test: 엔드포인트 별로 테스트 분리
